--- fees/thales/abis.ts
@@ -1,4 +1,0 @@
-export const OVERTIME_EVENT_ABI = {
-  safeboxFeePaid: "event SafeBoxFeePaid(uint safeBoxFee, uint safeBoxAmount, address collateral)",
-  safeboxSharePaid: "event SafeBoxSharePaid(uint safeBoxShare, uint safeBoxAmount)"
-};
