--- fees/thales/eventArgs.ts
@@ -1,10 +1,0 @@
-export interface ISafeBoxFeePaidEvent {
-  safeBoxFee: bigint;
-  safeBoxAmount: bigint;
-  collateral: string;
-}
-
-export interface ISafeBoxSharePaidEvent {
-  safeBoxShare: bigint;
-  safeBoxAmount: bigint;
-}
